Pass file and root directory hints to parse tasks.
This is needed for the meta-languages, which are configured via a config
file in the root directory. Mainly, it enables the meta-languages to
create explicit dependencies to tasks that generate source files, which
is needed to avoid hidden dependencies.

Concretely, this fixes an issue in ESV where opening an editor in the
IDE of an unpacked ESV file from libspoofax2 (which is unpacked with a
PIE task) would fail due to a hidden dependency. ESV now creates an
explicit dependency to the unarchive task in its parse task, which is
only now possible due to having access to the root directory to get
the configuration.

To achieve this, JSGLR1ParseTaskDef now takes a JSGLR1ParseTaskInput as
input, which contains an optional file and root directory hint.

Because of this, ConstraintAnalyzeMultiTaskDef now takes a function
as input that produces a mapping from file to AST supplier, to keep the
task decoupled from the type of parse inputs.