Improve "language management" and make it cross-platform.
"Language management" was responsible for dynamic loading and
composing multiple languages, but it was implemented in a rather
hacky way, and only worked in Eclipse.

This commit improves the existing "language management" and makes
it cross-platform. We've also renamed it to "component management",
since it is not just about languages, but also about managing Dagger
components, and even software components in general.

In `spoofax.core`, there is now a `mb.spoofax.core.component` package
that contains the component management functionality. Components
that wish to participant in component management need to implement the
`Participant` interface. This interface is used in:

- `StaticComponentManagerBuilder` which statically loads components.
  That is, those components need to be on the Java classpath of the
  current JVM.
- `DynamicComponentManager` which dynamically loads components. That
  is, those components can be loaded through compiled Java sources and
  resources.

Once loaded, the `Component` interface holds access to the
`ResourceServiceComponent`, `LanguageComponent` (optional),
`PieComponent`, and any subcomponents of the component.

Components are identified by `Coordinate`s, which use the conventional
GAV (group ID, artifact ID, version) used in the Java ecosystem.

Participants can specify dependencies to other components by
implementing the `getDependencies` method, allowing access to other
components through the `ComponentDependencyResolver` object in other
methods.

Participants can specify that their component would like to be composed
with other components by implementing the `getCompositionGroup` method.
Components that belong to the same composition group are composed,
meaning that their resource registries are composed into the same
`ResourceServiceComponent`, and their PIE tasks composed into the same
`PieComponent`.

Spoofax 3 now generates a participant implementation for each language:
`<Lang>Participant`, for example `Sdf3Participant`.

However, components do not have to be languages. For example, the
language workbench infrastructure implements `Participant` to compose
PIE tasks:

- `SpoofaxCompilerParticipant`
- `SpoofaxLwbCompilerParticipant`
- `DynamicLoadingParticipant`

The `Participant` interface is generic over the type of
`LoggerComponent`, `ResourceServiceComponent`, and `PlatformComponent`,
to allow platforms such as Eclipse to require superinterfaces for these
types through`EclipseParticipant`. As an example, the language workbench
infrastructure implements the following Eclipse participants:

- `SpoofaxCompilerEclipseParticipant`
- `SpoofaxLwbCompilerEclipseParticipant`
- `SpoofaxDynamicLoadingEclipseParticipant`
- `SpoofaxLwbEclipseParticipant`

The component management system is optional if the component does not
need to be composed or aware of other components, and does not need
to be dynamically loaded. In that case, the components can just be
manually instantiated. Currently, the CLI and IntelliJ projects of
languages do not use the component management system.

Additionally, the following smaller changes were made:

- Merge `spoofax.compiler.dagger` into `spoofax.compiler`.
- Merge `spoofax.lwb.compiler.dagger` into `spoofax.lwb.compiler`.
- Rename `Spoofax3` -> `SpoofaxLwb` in `spoofax.lwb.compiler`.
- Update org.jetbrains.intellij Gradle plugin to 1.0.