Fix transitive dependencies not set correctly
Projects with the constraint.common dependency didn't automatically
get the nabl2.terms and metaborg.util dependencies. This caused issues
in projects such as statix.intellij, where the newer version of Dagger
could not find the mb.nabl2.terms.stratego.StrategoTerms class.
Somehow this was not a problem in statix.eclipse because the
dependency was implicitly included through the bundleEmbedApi
configuration (from which the compileRuntime configuration derives).
This was also not an issue in statix.cli because its Dagger doesn't
build a StatixCliComponent that depends on a StatixModule.